Feature test

+++ app/controllers/prototypes_controller.rb
@@ -1,6 +1,7 @@
 class PrototypesController < ApplicationController
   before_action :authenticate_user!, except: [:index]
   before_action :set_prototype, only: [:show, :destroy, :edit, :update]
+
   def index
     @prototypes = Prototype.includes(:user).order("created_at DESC").page(params[:page]).per(5)
   end
@@ -12,10 +13,12 @@ class PrototypesController < ApplicationController
   end
 
   def destroy
-    if @prototype.destroy
-      redirect_to root_path , notice: "Prototype was successfully destroyed"
-    else
-      redirect_to action: :edit, alert: "Destroying was rejected"
+    if @prototype.user_id == current_user.id
+      if @prototype.destroy
+        redirect_to action: :index , notice: "Prototype was successfully destroyed"
+      else
+        redirect_to action: :edit, alert: "Destroying was rejected"
+      end
     end
   end
 
@@ -23,19 +26,19 @@ class PrototypesController < ApplicationController
   end
 
   def update
-    if @prototype.update(prototype_params)
-      redirect_to root_path, notice: "Prototype was successfully updated"
-    else
-      redirect_to edit_prototype_path(@prototype), alert: "Updating was rejected"
+    if @prototype.user_id == current_user.id
+      if @prototype.update(prototype_params)
+        redirect_to action: :index, notice: "Prototype was successfully updated"
+      else
+        redirect_to action: :edit, alert: "Updating was rejected"
+      end
     end
   end
 
   def new
     @prototype = Prototype.new
     @main_image = @prototype.prototype_images.build
-    @sub_images = []
-    3.times{ @sub_images << (@prototype.prototype_images.build) }
-    # 前までの書き方（@sub_images = 3.times {@prototype.prototype_images.build}）では@sub_imagesに代入されているのは　integer型の３だったためテストがうまくいかなかった。。rspecのおかげでミスに気づけた。ありがとうrspec
+    @sub_images = 3.times{ @prototype.prototype_images.build }
   end
 
   def create
@@ -43,7 +46,7 @@ class PrototypesController < ApplicationController
     if @prototype.save
       redirect_to root_path, notice: "Prototype was successfully created"
     else
-      redirect_to new_prototype_path, alert: "Posting was rejected"
+      redirect_to new_user_prototype(current_user), alert: "Posting was rejected"
     end
   end
